+++ tests/inputs-sort-requires/basic.lua
@@ -1,0 +1,4 @@
+local bee = require("b")
+local ah = require("a")
+
+print("hello world")