+++ tests/inputs-sort-requires/chained-dependency.lua
@@ -1,0 +1,6 @@
+local holder = script.Parent
+local modules = holder.Modules
+local cee = require(modules.Script)
+local bee = require(modules.BeeMovie)
+local aa = require(modules.Test)
+print("hello!")