+++ tests/inputs-sort-requires/groups.lua
@@ -1,0 +1,13 @@
+local ReplicatedStorage = game:GetService("ReplicatedStorage")
+local mainB = require(ReplicatedStorage.B)
+local mainA = require(ReplicatedStorage.A)
+
+local Packages = ReplicatedStorage.Packages
+local Z = require(Packages.Z)
+local Y = require(Packages.Y)
+local X = require(Packages.X)
+
+local Modules = ReplicatedStorage.Modules
+local C = require(Modules.C)
+local B = require(Modules.B)
+local A = require(Modules.A)