+++ tests/inputs-sort-requires/ignore-multi-require.lua
@@ -1,0 +1,2 @@
+local c = require("c")
+local b, a = require("b"), require("a")