+++ tests/inputs-sort-requires/separated-by-comments.lua
@@ -1,0 +1,9 @@
+-- Services
+local C = require("C")
+local B = require("B")
+local A = require("A")
+
+-- Packages
+local Z = require("Z")
+local Y = require("Y")
+local X = require("X")