+++ tests/inputs-sort-requires/sort-services-2.lua
@@ -1,0 +1,5 @@
+-- Requires should be treated as a separate group to services
+local ReplicatedStorage = game:GetService("ReplicatedStorage")
+local CollectionService = game:GetService("CollectionService")
+local C = require("C")
+local A = require("A")